requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.10.0
Bottleneck==1.3.2
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
cycler==0.11.0
Cython @ file:///tmp/build/80754af9/cython_1636023774860/work
docopt==0.6.2
filelock==3.4.0
flatbuffers==1.12
fonttools==4.28.3
gast==0.3.3
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.11.1
importlib-metadata==4.8.2
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.2
Keras==2.4.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
lvis==0.5.3
Markdown==3.3.6
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
networkx==2.6.3
numexpr @ file:///tmp/build/80754af9/numexpr_1618856527787/work
numpy==1.21.6
oauthlib==3.1.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.9.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.2
setuptools-scm==6.3.2
six==1.15.0
soupsieve==2.3.1
spatial-correlation-sampler==0.4.0
tb-nightly==2.8.0a20211117
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2021.11.2
tikzplotlib==0.9.15
tomli==1.2.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.9.1+cu111
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
typing-extensions==3.7.4.3
urllib3==1.26.7
visdom==0.1.8.9
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.12.1
yarg==0.1.9
zipp==3.6.0
